requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.10
pydantic-settings==2.11.0
bcrypt==5.0.0
pyjwt==2.10.1
python-multipart==0.0.20
httpx==0.28.1
websockets==15.0.1
google-auth==2.41.1
google-auth-httplib2==0.2.0
requests==2.32.5
redis==5.0.1
celery==5.3.4
alembic
asyncpg
fastapi
httpx
pyjwt
python-dotenv
python-multipart
sqlalchemy[asyncio]
uvicorn[standard]
websockets
alembic
pytest
pytest-asyncio
replit-object-storage
google-cloud-storage